requirements.txt.txt
fastapi
uvicorn
pydantic
numpy
scipy
matplotlib
Pillow
starlette
